Use Grid block as Post Template
