+++ src/blocks.js
@@ -9,4 +9,3 @@ import './blocks/button-container/block.js';
 import './blocks/button/block.js';
 import './blocks/headline/block.js';
 import './shared/style-imports.js';
-import './blocks/post-template';
