--- src/blocks/post-template/block.json
@@ -1,18 +1,0 @@
-{
-  	"$schema": "https://schemas.wp.org/trunk/block.json",
-  	"apiVersion": 2,
-  	"name": "generateblocks/post-template",
-  	"title": "Post Template",
-  	"category": "generateblocks",
-  	"parent": [ "generateblocks/query-loop" ],
-  	"usesContext": [
-	  	"generateblocks/query",
-		"generateblocks/gridId"
-	],
-  	"supports": {
-		"reusable": false,
-		"html": false,
-		"align": false,
-		"customClassName": false
-  	}
-}
