+++ src/blocks/query-loop/attributes.js
@@ -1,9 +1,9 @@
 import gridAttributes from '../grid/attributes';
 
 export default Object.assign( {}, gridAttributes, {
-	isQueryLoop: {
-		type: 'boolean',
-		default: false,
+	uniqueId: {
+		type: 'string',
+		default: '',
 	},
 
 	query: {
