+++ src/blocks/query-loop/components/PostTemplateRenderer.js
@@ -1,10 +1,8 @@
-import { normalizeRepeatableArgs, removeEmpty } from './utils';
-import { useSelect } from '@wordpress/data';
-import { store as coreStore } from '@wordpress/core-data';
 import { useMemo } from '@wordpress/element';
+import { useSelect } from '@wordpress/data';
 import LoopRenderer from './LoopRenderer';
-import { useBlockProps } from '@wordpress/block-editor';
-import classnames from 'classnames';
+import { normalizeRepeatableArgs, removeEmpty } from './utils';
+import { store as coreStore } from '@wordpress/core-data';
 
 const TEMPLATE = [
 	[ 'generateblocks/container', {
@@ -27,10 +25,9 @@ const TEMPLATE = [
 	] ],
 ];
 
-export default function PostTemplateEdit( props ) {
+export default function QueryLoopRenderer( props ) {
 	const { clientId, context } = props;
 	const query = context[ 'generateblocks/query' ] || {};
-	const gridId = context[ 'generateblocks/gridId' ];
 
 	const innerBlocks = useSelect( ( select ) => {
 		return select( 'core/block-editor' )?.getBlocks( clientId );
@@ -56,32 +53,21 @@ export default function PostTemplateEdit( props ) {
 		};
 	}, [ normalizedQuery ] );
 
-	const blockProps = useBlockProps( {
-		className: classnames( {
-			'gb-post-template': true,
-			[ `gb-post-template-${ gridId }` ]: true,
-		} ),
-	} );
-
 	return (
-		<div
-			{ ...blockProps }
-		>
-			<div className="gb-post-template-wrapper">
-				<LoopRenderer
-					data={ data }
-					hasData={ !! ( hasResolvedData && data?.length ) }
-					isResolvingData={ isResolvingData }
-					hasResolvedData={ hasResolvedData }
-					templateLock={ true }
-					template={ TEMPLATE }
-					innerBlocks={ innerBlocks }
-					contextCallback={ ( post ) => ( {
-						postType: post.type,
-						postId: post.id,
-					} ) }
-				/>
-			</div>
+		<div className="gb-post-template-wrapper">
+			<LoopRenderer
+				data={ data }
+				hasData={ !! ( hasResolvedData && data?.length ) }
+				isResolvingData={ isResolvingData }
+				hasResolvedData={ hasResolvedData }
+				templateLock={ true }
+				template={ TEMPLATE }
+				innerBlocks={ innerBlocks }
+				contextCallback={ ( post ) => ( {
+					postType: post.type,
+					postId: post.id,
+				} ) }
+			/>
 		</div>
 	);
 }
