+++ src/blocks/query-loop/editor.scss
@@ -1,9 +1,3 @@
-.wp-block-generateblocks-query-loop {
-    > .block-editor-inner-blocks > .block-editor-block-list__layout {
-        display: block; /* Un-flex standard Grid block */
-    }
-}
-
 .gb-post-template {
 	.gb-post-template-wrapper {
 		position: relative;
