+++ src/shared/editor.scss
@@ -129,3 +129,16 @@
 .block-editor-list-view-leaf.is-selected .gblocks-block-icon path {
     fill: currentColor !important;
 }
+
+.gblocks-action-panel {
+    padding: 0 !important;
+
+	.components-panel__body-title {
+		margin: 0 !important;
+	}
+
+	.components-panel__arrow path {
+		transform: rotate(90deg);
+		transform-origin: center;
+	}
+}
